[Added] the component for the events and moved the code for the movies to its own component
